+++ engine-tests/src/tests/sanity.rs
@@ -684,7 +684,7 @@ fn test_num_wasm_functions() {
     let module = walrus::ModuleConfig::default()
         .parse(runner.code.code())
         .unwrap();
-    let expected_number = 1494;
+    let expected_number = 1523;
     let actual_number = module.funcs.iter().count();
 
     assert!(