+++ engine-workspace/src/contract.rs
@@ -5,15 +5,16 @@ use crate::operation::{
     CallDeployCode, CallDeployErc20Token, CallDeployUpgrade, CallDeposit,
     CallFactorySetWNearAddress, CallFactoryUpdate, CallFactoryUpdateAddressVersion,
     CallFtOnTransfer, CallFtTransfer, CallFtTransferCall, CallFundXccSubAccount, CallMintAccount,
-    CallNew, CallNewEthConnector, CallPauseContract, CallPausePrecompiles, CallRefundOnError,
-    CallRegisterRelayer, CallRemoveEntryFromWhitelist, CallRemoveRelayerKey, CallResumeContract,
-    CallResumePrecompiles, CallSetEthConnectorContractAccount, CallSetEthConnectorContractData,
-    CallSetFixedGasCost, CallSetKeyManager, CallSetPausedFlags, CallSetSiloParams,
-    CallSetWhitelistStatus, CallStageUpgrade, CallStateMigration, CallStorageDeposit,
-    CallStorageUnregister, CallStorageWithdraw, CallSubmit, CallWithdraw, ViewAccountsCounter,
-    ViewBalance, ViewBlockHash, ViewBridgeProver, ViewChainId, ViewCode, ViewErc20FromNep141,
-    ViewFactoryWnearAddress, ViewFtBalanceOf, ViewFtBalanceOfEth, ViewFtMetadata,
-    ViewFtTotalEthSupplyOnAurora, ViewFtTotalEthSupplyOnNear, ViewFtTotalSupply,
+    CallMirrorErc20Token, CallNew, CallNewEthConnector, CallPauseContract, CallPausePrecompiles,
+    CallRefundOnError, CallRegisterRelayer, CallRemoveEntryFromWhitelist, CallRemoveRelayerKey,
+    CallResumeContract, CallResumePrecompiles, CallSetErc20Metadata,
+    CallSetEthConnectorContractAccount, CallSetEthConnectorContractData, CallSetFixedGasCost,
+    CallSetKeyManager, CallSetPausedFlags, CallSetSiloParams, CallSetWhitelistStatus,
+    CallStageUpgrade, CallStateMigration, CallStorageDeposit, CallStorageUnregister,
+    CallStorageWithdraw, CallSubmit, CallWithdraw, ViewAccountsCounter, ViewBalance, ViewBlockHash,
+    ViewBridgeProver, ViewChainId, ViewCode, ViewErc20FromNep141, ViewFactoryWnearAddress,
+    ViewFtBalanceOf, ViewFtBalanceOfEth, ViewFtMetadata, ViewFtTotalEthSupplyOnAurora,
+    ViewFtTotalEthSupplyOnNear, ViewFtTotalSupply, ViewGetErc20Metadata,
     ViewGetEthConnectorContractAccount, ViewGetFixedGasCost, ViewGetSiloParams,
     ViewGetWhitelistStatus, ViewIsUsedProof, ViewNep141FromErc20, ViewNonce, ViewOwner,
     ViewPausedFlags, ViewPausedPrecompiles, ViewStorageAt, ViewStorageBalanceOf, ViewUpgradeIndex,
@@ -22,8 +23,8 @@ use crate::operation::{
 use crate::transaction::{CallTransaction, ViewTransaction};
 use aurora_engine_types::account_id::AccountId;
 use aurora_engine_types::parameters::connector::{
-    FungibleTokenMetadata, PausedMask, Proof, SetEthConnectorContractAccountArgs,
-    WithdrawSerializeType,
+    Erc20Identifier, FungibleTokenMetadata, MirrorErc20TokenArgs, PausedMask, Proof,
+    SetErc20MetadataArgs, SetEthConnectorContractAccountArgs, WithdrawSerializeType,
 };
 use aurora_engine_types::parameters::engine::{
     CallArgs, FunctionCallArgsV2, NewCallArgs, NewCallArgsV2, RelayerKeyArgs, RelayerKeyManagerArgs,
@@ -210,6 +211,10 @@ impl EngineContract {
         CallDeployErc20Token::call(&self.contract).args_borsh(account_id)
     }
 
+    pub fn mirror_erc20_token(&self, args: MirrorErc20TokenArgs) -> CallMirrorErc20Token {
+        CallMirrorErc20Token::call(&self.contract).args_borsh(args)
+    }
+
     pub fn call(&self, contract: Address, amount: U256, input: Vec<u8>) -> CallCall {
         let value = WeiU256::from(amount);
         let args = CallArgs::V2(FunctionCallArgsV2 {
@@ -343,6 +348,10 @@ impl EngineContract {
     ) -> CallRemoveEntryFromWhitelist {
         CallRemoveEntryFromWhitelist::call(&self.contract).args_borsh(entry)
     }
+
+    pub fn set_erc20_metadata(&self, metadata: SetErc20MetadataArgs) -> CallSetErc20Metadata {
+        CallSetErc20Metadata::call(&self.contract).args_json(metadata)
+    }
 }
 
 /// View functions
@@ -470,6 +479,10 @@ impl EngineContract {
 
     pub fn factory_get_wnear_address(&self) -> ViewFactoryWnearAddress {
         ViewFactoryWnearAddress::view(&self.contract)
+    }
+
+    pub fn get_erc20_metadata(&self, identifier: Erc20Identifier) -> ViewGetErc20Metadata {
+        ViewGetErc20Metadata::view(&self.contract).args_json(identifier)
     }
 }